Conflicts:
	map-ear/target/m2e-wtp/ear-resources/META-INF/MANIFEST.MF
	map-ear/target/m2e-wtp/ear-resources/META-INF/maven/tn.esprit.pidev/map-ear/pom.properties
	map-ejb/target/classes/META-INF/MANIFEST.MF
	map-ejb/target/classes/META-INF/maven/tn.esprit.pidev/map-ejb/pom.properties
	map-web/target/m2e-wtp/web-resources/META-INF/MANIFEST.MF

+++ map-web/src/main/java/resources/MessageRessource.java
@@ -20,7 +20,10 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 
+import entities.Client;
 import entities.Message;
+import interfaces.ClientServiceLocale;
+import interfaces.ClientServiceRemote;
 import interfaces.messageRemote;
 
 @Path("mes")
@@ -29,6 +32,10 @@ public class MessageRessource {
 
 	@EJB(beanName="MessageService")
 	messageRemote em;
+
+
+	@EJB(beanName="ClientService")
+	ClientServiceLocale cl;
 	
 @POST
 @Consumes(MediaType.APPLICATION_JSON)
@@ -42,33 +49,51 @@ public Response add(Message m) {
 	}
 	else{
 	 int res = em.ajouterMessage(m);
-		return Response.status(Status.CREATED).entity("Created"+res).build();}
+		return Response.status(Status.CREATED).entity("Created"+m.getMessageId()).build();}
 	
 	}
 
 @PUT
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
-public Response Modifier(Message m) {
-	 int res =  em.modifier(m);
-		return Response.status(Status.OK).entity("Updated"+res).build();
-	}
+
+	public Response Modifier(Message m) {
+		if (!em.verif(m.getObject())||!em.verif(m.getType())||em.getMessagebyid(m.getMessageId())==null){
+			return Response.status(Status.NOT_ACCEPTABLE).entity("Verifier les Donnees").build();
+
+		}
+			else{
+
+		 int res =  em.modifier(m);
+			return Response.status(Status.OK).entity("Updated"+res).build();}
+		}
 @DELETE
 @Path("{id}")
 @Produces(MediaType.APPLICATION_JSON)
 public Response supprimer(@PathParam(value="id")int idmessage) {
-	   Message ms = em.getMessagebyid(idmessage);
+	if (em.getMessagebyid(idmessage)==null){
+		return Response.status(Status.NOT_ACCEPTABLE).entity("Verifier les Donnees").build();
+
+	}
+	else{
+	Message ms = em.getMessagebyid(idmessage);
 	   
      int res = em.delete(ms);
-		return Response.status(Status.OK).entity("Removed"+res).build();
+		return Response.status(Status.OK).entity("Removed"+res).build();}
 	}
 @GET
 @Path("{messageid}")
 @Produces(MediaType.APPLICATION_JSON)
 public Response getmessagebyid(@PathParam(value="messageid")int idmessage) {
 	 Message res =   em.getMessagebyid(idmessage);
-		return Response.status(Status.OK).entity("L element est"+res.getMessageId()+ "type est "+res.getType()).build();
+	
+	 if ( res== null ){ return Response.status(Status.BAD_REQUEST).entity("Verifier les Donnees").build();
+	 
 	}
+	 else{return Response.status(Status.OK).entity("L element est"+res.getMessageId()+ "type est "+res.getType()).build();
+		
+}	 
+}
 @GET
 @Path("getall")
 @Produces(MediaType.APPLICATION_JSON)
@@ -95,6 +120,34 @@ public Response getAllbyObject(@QueryParam("getbyobject") String Objet) {
 		return Response.status(Status.OK).entity(mes).build();
 	}
 
+@GET
+@Path("send")
+@Produces(MediaType.APPLICATION_JSON)
+public Response getmessagebysend(@QueryParam("idsend") int idmessage) {
+	List<Message> mes = new ArrayList<>();
+	mes=  em.getMessagebysend(idmessage);
+		return Response.status(Status.OK).entity(mes).build();
+		
+}	 
+@GET
+@Path("allclientsend")
+@Produces(MediaType.APPLICATION_JSON)
+public Response getmessagebyclient(@QueryParam("idclient") int client) {
+	List<Client> mes = new ArrayList<>();
+	mes=  em.getMessagebyClient(client);
+		return Response.status(Status.OK).entity(mes).build();
+		
+}
+@GET
+@Path("recu")
+@Produces(MediaType.APPLICATION_JSON)
+public Response getmessagebyrecu(@QueryParam("idrecu") int idmessage) {
+	List<Message> mes = new ArrayList<>();
+	mes=  em.getMessagebyrecu(idmessage);
+		return Response.status(Status.OK).entity(mes).build();
+		
+}
 
 
 }
+
